refactor: data,offscreen,sidepanel
