--- src/bo/index.d.ts
@@ -1,1 +1,0 @@
-declare module "@/bo/searchJobBO.js";
