+++ src/data/dto/statisticJobBrowseDTO.js
