+++ src/sidepanel/App.vue
