+++ src/sidepanel/main.js
@@ -3,7 +3,7 @@ import ElementPlus from 'element-plus'
 import 'element-plus/dist/index.css'
 import sidepanel from './App.vue'
 import  router  from "./router";
-import { initBridge } from '../../api/common';
+import { initBridge } from '@/api/common';
 import * as ElementPlusIconsVue from '@element-plus/icons-vue'
 
 async function init(){
